requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.14.2
selenium==4.11.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==2.2.5
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==2.1.5
pytz==2025.2
requests==2.31.0
six==1.17.0
soupsieve==2.4.1
typing_extensions==4.7.1
tzlocal==5.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
webdriver-manager==3.8.6
pyyaml==6.0.1